PostView2: Added xpltFileReader, the master file reader for xplt format.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@8913 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostViewLib/PltArchive.h
@@ -154,13 +154,13 @@ class OLeaf<const char*> : public OChunk
 public:
 	OLeaf(unsigned int nid, const char* sz) : OChunk(nid)
 	{
-		int l = strlen(sz);
+		int l = (int)strlen(sz);
 		m_psz = new char[l+1];
 		memcpy(m_psz, sz, l+1);
 	}
 	~OLeaf() { delete m_psz; }
 
-	int Size() { return strlen(m_psz)+sizeof(int); }
+	int Size() { return (int)strlen(m_psz)+sizeof(int); }
 	void Write(FileStream* fp)
 	{
 		fp->Write(&m_nID , sizeof(unsigned int), 1);
